Slimmed down many files, created singular Races file instead of one for each race, added skeleton of adding weapons, moved file structure, and small general fixes.
