+++ src/app/Character/Enums/level.ts
