+++ src/app/Character/Weapons/elemental.ts
@@ -1,5 +1,5 @@
 export interface Elemental {
   type: string;
   attack: number;
-  numberOfAttack: number;
+  numberOfAttacks: number;
 }
