--- src/app/Character/Weapons/rangeWeapon.ts
@@ -1,6 +1,0 @@
-import { Weapon } from './weapon';
-
-export class RangedWeapon extends Weapon {
-  range: number;
-  ammo?: number;
-}
