+++ src/app/Character/Weapons/weapon.ts
@@ -5,8 +5,13 @@ export class Weapon {
   attack: number;               // Number of sides the damage dice has
   numberOfAttacks: number;      // For the in case you have multiple attack dice
   critRange: number[];          // Values for which rolling a crit leads to crit damage
-  cirtDamage: number;           // Crit damage multiplier
-  description: string;
+  critDamage: number;           // Crit damage multiplier
+  type: string;
   modifier: string;
+  range: number;
   element?: Elemental;
+}
+
+export class RangedWeapon extends Weapon {
+  ammo?: number;
 }
